Refactored roles
