+++ include/roboteam_ai/stp/roles/Formation.h
