+++ src/stp/new_plays/GenericPass.cpp
@@ -6,16 +6,16 @@
 
 #include <roboteam_utils/Grid.h>
 #include <roboteam_utils/Tube.h>
-#include <stp/new_roles/Formation.h>
+#include <stp/roles/Formation.h>
 
 #include "stp/invariants/BallClosestToUsInvariant.h"
 #include "stp/invariants/BallOnOurSideInvariant.h"
 #include "stp/invariants/FreedomOfRobotsInvariant.h"
 #include "stp/invariants/game_states/NormalPlayGameStateInvariant.h"
-#include "stp/new_roles/Halt.h"
-#include "stp/new_roles/Keeper.h"
-#include "stp/new_roles/PassReceiver.h"
-#include "stp/new_roles/Passer.h"
+#include "stp/roles/Halt.h"
+#include "stp/roles/Keeper.h"
+#include "stp/roles/PassReceiver.h"
+#include "stp/roles/Passer.h"
 
 namespace rtt::ai::stp::play {
