+++ src/stp/new_plays/Halt.cpp
@@ -5,7 +5,7 @@
 #include "stp/new_plays/Halt.h"
 
 #include "stp/invariants/game_states/HaltGameStateInvariant.h"
-#include "stp/new_roles/Halt.h"
+#include "stp/roles/Halt.h"
 
 namespace rtt::ai::stp::play {
