+++ src/stp/new_plays/TimeOut.cpp
@@ -5,7 +5,7 @@
 #include "stp/new_plays/TimeOut.h"
 
 #include "stp/invariants/game_states/TimeOutGameStateInvariant.h"
-#include "stp/new_roles/Formation.h"
+#include "stp/roles/Formation.h"
 
 namespace rtt::ai::stp::play {
