+++ src/stp/roles/Attacker.cpp
@@ -2,7 +2,7 @@
 // Created by jordi on 17-03-20.
 //
 
-#include "stp/new_roles/Attacker.h"
+#include "stp/roles/Attacker.h"
 
 #include "stp/tactics/GetBall.h"
 #include "stp/tactics/KickAtPos.h"
@@ -13,5 +13,4 @@ Attacker::Attacker(std::string name) : Role(std::move(name)) {
     // create state machine and initializes the first state
     robotTactics = collections::state_machine<Tactic, Status, StpInfo>{tactic::GetBall(), tactic::KickAtPos()};
 }
-
 }  // namespace rtt::ai::stp::role
