+++ src/stp/roles/BallPlacer.cpp
@@ -2,7 +2,7 @@
 // Created by jesse on 07-04-20.
 //
 
-#include "stp/new_roles/BallPlacer.h"
+#include "stp/roles/BallPlacer.h"
 
 #include "stp/tactics/AvoidBall.h"
 #include "stp/tactics/DriveWithBall.h"
@@ -14,5 +14,4 @@ BallPlacer::BallPlacer(std::string name) : Role(std::move(name)) {
     // create state machine and initializes the first state
     robotTactics = collections::state_machine<Tactic, Status, StpInfo>{tactic::GetBall(), tactic::DriveWithBall(), tactic::AvoidBall()};
 }
-
 }  // namespace rtt::ai::stp::role
